Feature option to configure different URLs for redis cache
This feature will address #1420 issue, allowing users to set individual `REDIS_URL` for Geocoder, Sidekiq, and Websocket.

Three additional optional environment variables are introduced in this commit, namely:
 - `REDIS_GEOCODER_URL`
 - `REDIS_SIDEKIQ_URL`
 - `REDIS_WS_URL`

These environment variables will takes priority over `REDIS_URL` env variable if set.

There is a possibility for a potential bug where the user can accidentally set the same values for two of the mentioned env variables. I did not implement the condition to check this scenario since it could complicate the code, but can implement if needed.

This is my first contribution to ruby code, so corrections and guidance are welcomed.